little arch refactoring
