+++ app/src/main/java/com/example/playlistmaker/data/storage/SharedPrefNightModeStorage.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import com.example.playlistmaker.data.NightModeStorage
 import com.example.playlistmaker.data.storage.dto.NightModeDto
 
-class SharedPrefNightModeStorage(context: Context) :NightModeStorage {
+class SharedPrefNightModeStorage(context: Context) : NightModeStorage {
 
     private val sharedPreferences = context.getSharedPreferences(NIGHT_MODE, Context.MODE_PRIVATE)
     override fun save(isNight: NightModeDto) {
@@ -16,7 +16,7 @@ class SharedPrefNightModeStorage(context: Context) :NightModeStorage {
         return NightModeDto(isNight)
     }
 
-    private companion object{
+    private companion object {
         const val NIGHT_MODE = "night_mode"
         const val IS_NIGHT = "is_night"
         const val DEFAULT_NIGHT_MODE = false
